social media notifyier bot/requirements.txt
discord.py
python-dotenv
TikTokApi
google-api-python-client